+++ R/imports.R
@@ -5,6 +5,7 @@
 #' @importFrom taxize tnrs
 #' @importFrom foreign read.dbf
 #' @importFrom stringdist stringsim
+#' @importClassesFrom biblio lib_df
 # #' @importFrom taxa taxon_ids taxon_ranks
 # #' @importClassesFrom biblio lib_df
 NULL